+++ packages/contracts-rfq/test/integration/FastBridge.MulticallTarget.t.sol
@@ -5,7 +5,11 @@ import {FastBridge} from "../../contracts/FastBridge.sol";
 
 import {IFastBridge, MulticallTargetIntegrationTest} from "./MulticallTarget.t.sol";
 
+// solhint-disable no-empty-blocks
 contract FastBridgeMulticallTargetTest is MulticallTargetIntegrationTest {
+    /// @notice We include an empty "test" function so that this contract does not appear in the coverage report.
+    function testFastBridgeMulticallTargetTest() external {}
+
     function deployAndConfigureFastBridge() public override returns (address) {
         FastBridge fastBridge = new FastBridge(address(this));
         fastBridge.grantRole(fastBridge.RELAYER_ROLE(), relayer);