+++ packages/sdk-router/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.11.9](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.11.8...@synapsecns/sdk-router@0.11.9) (2024-12-18)
+
+**Note:** Version bump only for package @synapsecns/sdk-router
+
+
+
+
+
 ## [0.11.8](https://github.com/synapsecns/sanguine/compare/@synapsecns/sdk-router@0.11.7...@synapsecns/sdk-router@0.11.8) (2024-12-01)